rename sql generator to spring-data-relational
