+++ generators/generator-list.mjs
@@ -75,7 +75,7 @@ export const GENERATOR_SPRING_CACHE = 'spring-cache';
 export const GENERATOR_SPRING_CONTROLLER = 'spring-controller';
 export const GENERATOR_SPRING_SERVICE = 'spring-service';
 export const GENERATOR_SPRING_WEBSOCKET = 'spring-websocket';
-export const GENERATOR_SQL = 'sql';
+export const GENERATOR_SPRING_DATA_RELATIONAL = 'spring-data-relational';
 export const GENERATOR_UPGRADE = 'upgrade';
 export const GENERATOR_VUE = 'vue';
 export const GENERATOR_WORKSPACES = 'workspaces';
