+++ generators/spring-data-relational/cleanup.mts
