+++ generators/spring-data-relational/entity-files.mjs
