+++ generators/spring-data-relational/internal/dependencies.mts
