+++ generators/spring-data-relational/support/application-properties.mts
