+++ generators/spring-data-relational/templates/src/main/java/package/config/DatabaseConfiguration.java.ejs
