+++ generators/spring-data-relational/templates/src/main/java/package/domain/_PersistClass_.java.jhi.jakarta_lifecycle_events.ejs
