+++ generators/spring-data-relational/templates/src/main/java/package/domain/_PersistClass_Callback.java.ejs
