+++ generators/spring-data-relational/templates/src/main/java/package/repository/EntityManager_reactive.java.ejs
