+++ generators/spring-data-relational/templates/src/main/java/package/repository/UserSqlHelper_reactive.java.ejs
