+++ generators/spring-data-relational/templates/src/main/java/package/repository/_EntityClass_RepositoryInternalImpl_reactive.java.ejs
