+++ generators/spring-data-relational/templates/src/main/java/package/repository/_EntityClass_RepositoryWithBagRelationships.java.ejs
