+++ generators/spring-data-relational/templates/src/main/java/package/repository/_EntityClass_SqlHelper_reactive.java.ejs
