+++ generators/spring-data-relational/templates/src/main/java/package/repository/rowmapper/ColumnConverter_reactive.java.ejs
