+++ generators/spring-data-relational/templates/src/main/java/package/repository/rowmapper/UserRowMapper_reactive.java.ejs
