+++ generators/spring-data-relational/templates/src/main/java/package/repository/rowmapper/_EntityClass_RowMapper_reactive.java.ejs
