+++ generators/spring-data-relational/templates/src/main/resources/.h2.server.properties.ejs
