+++ generators/spring-data-relational/templates/src/test/java/package/config/EmbeddedSQL.java.ejs
