+++ generators/spring-data-relational/templates/src/test/java/package/config/MariadbTestContainer.java.ejs
