+++ generators/spring-data-relational/templates/src/test/java/package/config/MsSqlTestContainer.java.ejs
