+++ generators/spring-data-relational/templates/src/test/java/package/config/MysqlTestContainer.java.ejs
