+++ generators/spring-data-relational/templates/src/test/java/package/config/PostgreSqlTestContainer.java.ejs
