+++ generators/spring-data-relational/templates/src/test/java/package/config/SqlTestContainer.java.ejs
