+++ generators/spring-data-relational/templates/src/test/java/package/config/timezone/HibernateTimeZoneIT.java.ejs
