+++ generators/spring-data-relational/templates/src/test/java/package/repository/timezone/DateTimeWrapper.java.ejs
