+++ generators/spring-data-relational/templates/src/test/resources/config/application-testdev.yml.ejs
