+++ generators/spring-data-relational/templates/src/test/resources/config/application-testprod.yml.ejs
