+++ generators/spring-data-relational/templates/src/test/resources/testcontainers/mariadb/my.cnf
